Refactor: Organize TLS Extensions into TLS 1.2 and 1.3 Modules
Many extensions are understood and semanticaly relevant by both revisions
but some are strongly version specific. This moves the lattter ones into
their respective tls12 and tls13 modules. The selection roughly follows
RFC 8446 Section 4.2: extensions that aren't listed there (anymore) are
classified as specific to TLS 1.2, whereas extensions that are introduced
by RFC 8446 and that clearly aren't retrofittable, are classified as
specific to TLS 1.3.

If either TLS 1.2 or TLS 1.3 is disabled at build time, such extensions
are not parsed by Botan but their content is captured by the
Unknown_Extension class instead.
