+++ src/lib/tls/tls12/msg_client_hello_12.cpp
@@ -11,6 +11,7 @@
 
 #include <botan/tls_callbacks.h>
 #include <botan/tls_exceptn.h>
+#include <botan/tls_extensions_12.h>
 #include <botan/tls_policy.h>
 #include <botan/internal/stl_util.h>
 #include <botan/internal/tls_handshake_hash.h>
