+++ src/lib/tls/tls12/msg_server_hello_12.cpp
@@ -10,7 +10,7 @@
 #include <botan/tls_messages_12.h>
 
 #include <botan/tls_callbacks.h>
-#include <botan/tls_extensions.h>
+#include <botan/tls_extensions_12.h>
 #include <botan/tls_policy.h>
 #include <botan/internal/tls_handshake_hash.h>
 #include <botan/internal/tls_handshake_io.h>
