+++ src/lib/tls/tls12/tls_handshake_state.h
@@ -15,6 +15,7 @@
 #include <botan/tls_exceptn.h>
 #include <botan/tls_extensions.h>
 #include <botan/tls_handshake_msg.h>
+#include <botan/tls_session.h>
 #include <botan/internal/tls_handshake_hash.h>
 #include <botan/internal/tls_handshake_io.h>
 #include <botan/internal/tls_handshake_transitions.h>
