+++ src/lib/tls/tls13/info.txt
@@ -9,6 +9,7 @@ brief -> "TLS 1.3 protocol implementation"
 
 <header:public>
 tls_messages_13.h
+tls_extensions_13.h
 tls_psk_identity_13.h
 </header:public>
