+++ src/lib/tls/tls13/msg_certificate_req_13.cpp
@@ -12,6 +12,7 @@
 #include <botan/pkix_types.h>
 #include <botan/tls_callbacks.h>
 #include <botan/tls_exceptn.h>
+#include <botan/tls_extensions_13.h>
 #include <botan/tls_policy.h>
 #include <botan/internal/tls_reader.h>
