+++ src/lib/tls/tls13/msg_server_hello_13.cpp
@@ -13,6 +13,7 @@
 #include <botan/tls_alert.h>
 #include <botan/tls_callbacks.h>
 #include <botan/tls_exceptn.h>
+#include <botan/tls_extensions_13.h>
 #include <botan/tls_policy.h>
 #include <botan/internal/loadstor.h>
 #include <botan/internal/stl_util.h>
