+++ src/lib/tls/tls13/tls_extensions_psk.cpp
@@ -7,7 +7,7 @@
 * Botan is released under the Simplified BSD License (see license.txt)
 */
 
-#include <botan/tls_extensions.h>
+#include <botan/tls_extensions_13.h>
 
 #include <botan/credentials_manager.h>
 #include <botan/hash.h>
