+++ src/lib/tls/tls13/tls_server_impl_13.cpp
@@ -11,6 +11,7 @@
 #include <botan/credentials_manager.h>
 #include <botan/rng.h>
 #include <botan/tls_callbacks.h>
+#include <botan/tls_extensions_13.h>
 #include <botan/tls_policy.h>
 #include <botan/internal/stl_util.h>
 #include <botan/internal/tls_cipher_state.h>
