+++ src/lib/tls/tls_session.cpp
@@ -22,6 +22,7 @@
 #include <botan/internal/stl_util.h>
 
 #if defined(BOTAN_HAS_TLS_13)
+   #include <botan/tls_extensions_13.h>
    #include <botan/tls_messages_13.h>
 #endif
