+++ src/tests/test_tls_rfc8448.cpp
@@ -13,9 +13,9 @@
 
 // Since RFC 8448 uses a specific set of cipher suites we can only run this
 // test if all of them are enabled.
-#if defined(BOTAN_HAS_TLS_13) && defined(BOTAN_HAS_AEAD_CHACHA20_POLY1305) && defined(BOTAN_HAS_AEAD_GCM) &&          \
-   defined(BOTAN_HAS_AES) && defined(BOTAN_HAS_X25519) && defined(BOTAN_HAS_SHA2_32) && defined(BOTAN_HAS_SHA2_64) && \
-   defined(BOTAN_HAS_ECDSA) && defined(BOTAN_HAS_PSS)
+#if defined(BOTAN_HAS_TLS_12) && defined(BOTAN_HAS_TLS_13) && defined(BOTAN_HAS_AEAD_CHACHA20_POLY1305) &&             \
+   defined(BOTAN_HAS_AEAD_GCM) && defined(BOTAN_HAS_AES) && defined(BOTAN_HAS_X25519) && defined(BOTAN_HAS_SHA2_32) && \
+   defined(BOTAN_HAS_SHA2_64) && defined(BOTAN_HAS_ECDSA) && defined(BOTAN_HAS_PSS)
    #define BOTAN_CAN_RUN_TEST_TLS_RFC8448
 #endif
 
@@ -31,7 +31,8 @@
    #include <botan/pkcs8.h>
    #include <botan/tls_callbacks.h>
    #include <botan/tls_client.h>
-   #include <botan/tls_extensions.h>
+   #include <botan/tls_extensions_12.h>
+   #include <botan/tls_extensions_13.h>
    #include <botan/tls_messages.h>
    #include <botan/tls_policy.h>
    #include <botan/tls_server.h>
