+++ src/tests/unit_tls.cpp
@@ -21,6 +21,7 @@
    #include <botan/tls_client.h>
    #include <botan/tls_exceptn.h>
    #include <botan/tls_extensions.h>
+   #include <botan/tls_external_psk.h>
    #include <botan/tls_policy.h>
    #include <botan/tls_server.h>
    #include <botan/tls_session_manager_memory.h>
